Move formatter factory helper into source
