+++ envoy/formatter/BUILD
@@ -28,10 +28,12 @@ envoy_cc_library(
     ],
     deps = [
         ":http_formatter_context_interface",
+        "//envoy/common:pure_lib",
         "//envoy/config:typed_config_interface",
-        "//envoy/http:header_map_interface",
-        "//envoy/registry",
         "//envoy/server:factory_context_interface",
         "//envoy/stream_info:stream_info_interface",
+        "//source/common/protobuf",
+        "@abseil-cpp//absl/strings",
+        "@abseil-cpp//absl/types:optional",
     ],
 )
